Add flattening and non flattening options for recursive
The flattening option changes the structure of the resulting object.

When enabled (default) the object will be simple and its keys will be
the files name, independently of their location.

When disabled, the object will consist of nested objects where the keys
correspond to the relative path of the file compared to the starting
directory.